+++ docs/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project does not adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 This project uses MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH.
 
+## [1.16.5-4.0.0.7] - 2022.01.19
+### Added
+- Added `es_es.json` localization (thanks albertosaurio65!) [#53](https://github.com/TheIllusiveC4/CulinaryConstruct/pull/53)
+### Fixed
+- Fixed Suspicious Stew not applying effects when used as an ingredient [#51](https://github.com/TheIllusiveC4/CulinaryConstruct/issues/51)
+
 ## [1.16.5-4.0.0.6] - 2021.02.25
 ### Fixed
 - Fixed some bugs using container ingredient stacks [#49](https://github.com/TheIllusiveC4/CulinaryConstruct/issues/49)